Revert "Refactore Bento."
This reverts commit 52e3bf0d1698cf2b130f9a477ffd934b53d68329.
